requirements.txt
instagrapi==2.1.2
pillow==11.0.0
python-dotenv==1.0.1